Map and enemy arrays won't be cleared when disabling
The lines/monsters array will no longer be cleared when the minimap is disabled; it just won't be updated anymore. The main point here is if the player starts fiddling with this option and sets it to off and then back to on, they shouldn't be greeted with a blank minimap screen because they turned off the option and the arrays got cleared.